refactor: decrease size of package
