+++ .changeset/sour-carrots-fail.md
@@ -1,0 +1,5 @@
+---
+'generaltranslation': patch
+---
+
+refactor: decrease package size
